Move quick-start/setup.md to quickstart.md
In other words, I deleted an old `quickstart.md` and replaced it with
`quick-start/setup.md`.
The reason is that we can no longer maintain the repository
"sider/enterprise-config," which are referred by the old
`quickstart.md`.

I know this new `quickstart.md` is not "quick."
I want to elaborate Sider Enterprise itself as well as the document in
the future.