Fix confusing URLs (#435)
The URLs of PHP Code_Sniffer and Rails Best Practices are confusing
due to the mismatch between `sider.yml` and the document path, e.g. `linter.code_sniffer` and `tools/php/codesniffer`.

So, this change fixes the mismatch. In addition, it adds redirect rules to prevent the existing URLs from break.